playing around with getting milliseconds timestamp into mongodb

+++ dataset/GitHub_Java/Graylog2.graylog2-server/src/main/java/org/graylog2/messagehandlers/gelf/GELFMessage.java
@@ -33,6 +33,7 @@ import org.graylog2.streams.matchers.StreamRuleMatcherIF;
 import java.util.*;
 import java.util.regex.Pattern;
 import java.util.zip.Deflater;
+import org.json.simple.JSONObject;
 import org.json.simple.JSONValue;
 
 /**
@@ -57,7 +58,7 @@ public class GELFMessage {
     private Map<String, String> additionalData = new HashMap<String, String>();
     private List<Stream> streams = null;
     private boolean convertedFromSyslog = false;
-    private double createdAt = 0;
+    private float createdAt = 0;
 
     private boolean filterOut = false;
     private boolean doRouting = true;
@@ -439,14 +440,14 @@ public class GELFMessage {
     /**
      * @return the createdAt
      */
-    public double getCreatedAt() {
+    public float getCreatedAt() {
         return createdAt;
     }
 
     /**
      * @param createdAt the createdAt to set
      */
-    public void setCreatedAt(double createdAt) {
+    public void setCreatedAt(float createdAt) {
         this.createdAt = createdAt;
     }
